fix(ui): refactor and remove useSubworkCreate hook
